Make the strategyproof property of the discrete game EXPLICITLY.
Actually this revision corresponds to Reviewer 1's Comment 1:
``
So, it is better to give a lemma or a theorem =
to explicitly prove that the proposed mechanism is strategyproof. =
Actually, only Lemma 3 in this paper is related to this issue. =
Nevertheless, this Lemma still has not shown the strategyproof property =
explicitly.
''