This edit corresponds to the second item by Reviewer 1:
2) The model in this paper has some limitations. The authors did not =
consider some important issues in the platform, which are generally =
concerned in other papers, such as the number of tasks to be finished and =
the budget constraint of the platform to finish these tasks. When these =
issues are taken into consideration, the problem might become more =
difficult. Despite this, it is better to give an explanation or discussion =
on this, to make the proposed mechanism practical.

I have given some discussions on budget feasibility. That is, the sum of the payments to users can be upper bounded by the budget.

Note that the number of tasks in this item has not yet finished.